requirements.txt
python-telegram-bot==20.7
snscrape
python-dotenv
requests
telegram